Bug fixing for studentInClass
Problem 1
The bug is caused by the function array_search()
When the element is in the index, it will return the key of the
element, which sometimes could be 0, equivalent to false in weak equal.
The problem is solved by checking with the formula array_search() ===
false

Problem 2
The bug is caused by SAE’s erratic behavior. It cannot accept simply
using fetch_assoc()[‘table_column_name’].